+++ CHANGES.md
@@ -4,7 +4,7 @@ These are the changes brought in each versions:
 v1.0.6 (TBD)
 ===================
 Non-breaking changes:
-* Fix: Spaces or empty string accepted as valid Decimal values (interpreted as 0)
+* Fix: Spaces or empty string accepted as valid Integer and Decimal values (interpreted as 0)
 
 v1.0.5 (2020-09-16)
 ===================